+++ src/main/scala/lerna/akka/entityreplication/raft/RaftActor.scala
@@ -465,31 +465,34 @@ private[raft] class RaftActor(
     }
     commitLogStore ! CommitLogStoreActor.AppendCommittedEntries(shardId, Seq.empty)
 
-    if (
-      currentData.replicatedLog.entries.size >= settings.compactionLogSizeThreshold
-      && currentData.hasLogEntriesThatCanBeCompacted
-    ) {
+    if (currentData.replicatedLog.entries.size >= settings.compactionLogSizeThreshold) {
       val estimatedCompactedLogSize: Int =
         currentData.estimatedReplicatedLogSizeAfterCompaction(settings.compactionPreserveLogSize)
       if (estimatedCompactedLogSize >= settings.compactionLogSizeThreshold) {
+        // This warning might also happen when the first SnapshotTick expires if
+        //   * there are already enough entries (>= compactionLogSizeThreshold), and eventSourcingIndex is unknown yet.
+        //   * there is no leader (e.g., split vote), and no entries are applied.
+        //   * there is the leader, but only a few entries are applied (the replication is too slow).
         if (log.isWarningEnabled) {
           log.warning(
-            "[{}] Skipping compaction since compaction might not delete enough entries " +
-            "(even if this compaction continues, the remaining entries will trigger new compaction at the next tick). " +
+            "[{}] Compaction might not delete enough entries, but will continue to reduce log size as possible " +
+            "(even if this compaction continues, the remaining entries might trigger new compaction at the next tick). " +
             s"Estimated compacted log size is [{}] entries (lastApplied [{}], eventSourcingIndex [{}], preserveLogSize [${settings.compactionPreserveLogSize}]), " +
             s"however compaction.log-size-threshold is [${settings.compactionLogSizeThreshold}] entries. " +
-            "This warning happens if event sourcing is too slow or compaction is too fast.",
+            "This warning might happen if event sourcing is too slow or compaction is too fast (or too slow). " +
+            "If this warning continues, please consult settings related to event sourcing and compaction.",
             currentState,
             estimatedCompactedLogSize,
             currentData.lastApplied,
             currentData.eventSourcingIndex,
           )
         }
-      } else if (snapshotSynchronizationIsInProgress) {
+      }
+      if (snapshotSynchronizationIsInProgress) {
         // Snapshot updates during synchronizing snapshot will break consistency
         if (log.isInfoEnabled)
           log.info("Skipping compaction because snapshot synchronization is in progress")
-      } else {
+      } else if (currentData.hasAppliedLogEntries) {
         val (term, logEntryIndex, entityIds) = currentData.resolveSnapshotTargets()
         applyDomainEvent(SnapshottingStarted(term, logEntryIndex, entityIds)) { _ =>
           if (log.isInfoEnabled)